% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeze.R
\name{freeze}
\alias{freeze}
\title{Freeze installed packages from global environment}
\usage{
freeze(version = F, export = T)
}
\arguments{
\item{version}{Boolean. Show version or not}

\item{export}{Boolean. Save into RDS file or not}
}
\value{
Installed packages
}
\description{
Freeze installed packages from global environment
}
\examples{
freeze()
}
